% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.starray.R
\name{as.starray}
\alias{as.starray}
\alias{starray}
\alias{as_starray}
\title{Convert array to \code{starray}}
\usage{
as.starray(x)

starray(x)

as_starray(x)
}
\arguments{
\item{x}{A three-dimensional array}
}
\value{
An \code{starray} object.
}
\description{
Convert a three-dimensional spatio-temporal array into
an \code{starray} object.  The first two dimensions are
assumed to relate to gridded spatial positions.
}
\examples{
star = as.starray(tasmax)
class(star)
}
