% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulate.R
\name{circulate}
\alias{circulate}
\title{Circulate values of a vector}
\usage{
circulate(x, n = 1)
}
\arguments{
\item{x}{vector of values}

\item{n}{The number of values to circulate}
}
\value{
The circulated vector
}
\description{
The first \code{n} values of \code{x} are circulated
from the front of \code{x} to the back of \code{x}.
}
\examples{
circulate(1:10, n = 2)
circulate(as.list(1:10), n = 2)
}
\author{
Joshua French
}
