% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect \code{hero_radsplines}}
\usage{
connect(...)
}
\arguments{
\item{...}{A sequence of \code{hero_radspline} objects from the
\code{\link{radspline}} function.}
}
\value{
A combined \code{hero_radspline}
}
\description{
\code{connect} joins multiple \code{hero_radspline}
objects into a single \code{hero_radspline}.
The \code{e}
}
\examples{
border = border.grid(lon, lat)
s1 = radspline(nknots = 36, border = border)
plot(s1)
s2 = radspline(nknots = 36 * 4, border = border,
               width = 6)
plot(s2)
par(mfrow = c(1, 2))
plot(s1)
plot(s2)
par(mfrow = c(1, 1))
s = connect(s1, s2)
plot(s)
}
\seealso{
\code{\link{radspline}}
}
