% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffpen.R
\name{diffpen}
\alias{diffpen}
\title{P-spline difference penalty}
\usage{
diffpen(x, m = 2, sparse = TRUE)
}
\arguments{
\item{x}{A \code{hero_bspline} object produced by \code{\link{bspline}}.}

\item{m}{A positive integer indicating order of the difference penalty.}

\item{sparse}{A logical value indicating if the result
should be a sparse version of the
\code{\link[Matrix]{Matrix-class}}.}
}
\value{
A \code{\link[base]{matrix}} or \code{\link[Matrix]{sparseMatrix-class}} object.
}
\description{
P-spline difference penalty
}
\examples{
b = bspline(nbasis = 10)
diffpen(b)
}
\author{
Joshua French
}
