% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hero_enlarge.R
\name{plot.hero_enlarge}
\alias{plot.hero_enlarge}
\title{Plot a \code{hero_enlarge} object}
\usage{
\method{plot}{hero_enlarge}(x, ..., blist = list(col = "grey"))
}
\arguments{
\item{x}{An object of class \code{hero_enlarge}.}

\item{...}{Additional graphical parameters passed to the
plotting method for \code{\link[sp]{SpatialPolygons-class}}
for \code{x$eborder}}

\item{blist}{A list of additional graphical parameters
passed to the plotting method for \code{\link[sp]{SpatialPolygons-class}}
for \code{x$border}.}
}
\description{
Plot the enlarged and original border defined be a
set of coordinates.
}
\examples{
b = border.grid(lon, lat)
e = enlarge(b)
plot(e)
}
\seealso{
\code{\link[sp]{SpatialPolygons-class}}
}
