% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data for sandwich smooth}
\usage{
prepare(data, ...)
}
\arguments{
\item{data}{The data to prepare}

\item{...}{Not implemented}
}
\value{
A prepared object
}
\description{
A generic function to prepare various types of data.  See
the functions linked in See Also.
}
\seealso{
\code{\link{prepare.numeric}},
\code{\link{prepare.matrix}},
\code{\link{prepare.array}},
\code{\link{prepare.sts}},
\code{\link{prepare.starray}}
}
